+++ src/tooltip/test/tooltip.spec.js
@@ -132,6 +132,38 @@ describe('tooltip', function() {
     expect( elmBody.children().length ).toBe( 0 );
   }));
 
+  describe('with specified enable expression', function() {
+
+    beforeEach(inject(function ($compile) {
+      scope.enable = false;
+      elmBody = $compile(angular.element(
+        '<div><span tooltip="tooltip text" tooltip-enable="enable">Selector Text</span></div>'
+      ))(scope);
+      scope.$digest();
+      elm = elmBody.find('span');
+      elmScope = elm.scope();
+
+    }));
+
+    it('should not open ', inject(function () {
+
+      elm.trigger('mouseenter');
+      expect(elmScope.tt_isOpen).toBeFalsy();
+      expect(elmBody.children().length).toBe(1);
+
+    }));
+
+    it('should open', inject(function () {
+
+      scope.enable = true;
+      scope.$digest();
+      elm.trigger('mouseenter');
+      expect(elmScope.tt_isOpen).toBeTruthy();
+      expect(elmBody.children().length).toBe(2);
+
+    }));
+  });
+
   describe('with specified popup delay', function () {
 
     beforeEach(inject(function ($compile) {
@@ -208,6 +240,32 @@ describe('tooltip', function() {
       expect( elmScope.tt_isOpen ).toBeTruthy();
       elm.trigger('fakeTriggerAttr');
       expect( elmScope.tt_isOpen ).toBeFalsy();
+    }));
+
+    it('should not share triggers among different element instances - issue 692', inject( function ($compile) {
+
+      scope.test = true;
+      elmBody = angular.element(
+        '<div>' +
+          '<input tooltip="Hello!" tooltip-trigger="{{ (test && \'mouseenter\' || \'click\') }}" />' +
+          '<input tooltip="Hello!" tooltip-trigger="{{ (test && \'mouseenter\' || \'click\') }}" />' +
+        '</div>'
+      );
+
+      $compile(elmBody)(scope);
+      scope.$apply();
+      var elm1 = elmBody.find('input').eq(0);
+      var elm2 = elmBody.find('input').eq(1);
+      var elmScope1 = elm1.scope();
+      var elmScope2 = elm2.scope();
+
+      scope.$apply('test = false');
+
+      elm2.trigger('mouseenter');
+      expect( elmScope2.tt_isOpen ).toBeFalsy();
+
+      elm2.click();
+      expect( elmScope2.tt_isOpen ).toBeTruthy();
     }));
   });
 
@@ -335,6 +393,11 @@ describe( 'tooltipHtmlUnsafe', function() {
     scope.$digest();
     elm = elmBody.find('span');
     elmScope = elm.scope();
+  }));
+
+  it( 'should render html properly', inject( function () {
+    elm.trigger( 'mouseenter' );
+    expect( elmBody.find('.tooltip-inner').html() ).toBe( scope.html );
   }));
 
   it( 'should show on mouseenter and hide on mouseleave', inject( function () {
@@ -465,6 +528,24 @@ describe( '$tooltipProvider', function() {
         elm.trigger('focus');
         expect( elmScope.tt_isOpen ).toBeTruthy();
         elm.trigger('blur');
+        expect( elmScope.tt_isOpen ).toBeFalsy();
+      }));
+
+      it( 'should override the show and hide triggers if there is an attribute', inject( function ( $rootScope, $compile ) {
+        elmBody = angular.element(
+          '<div><input tooltip="tooltip text" tooltip-trigger="mouseenter"/></div>'
+        );
+
+        scope = $rootScope;
+        $compile(elmBody)(scope);
+        scope.$digest();
+        elm = elmBody.find('input');
+        elmScope = elm.scope();
+
+        expect( elmScope.tt_isOpen ).toBeFalsy();
+        elm.trigger('mouseenter');
+        expect( elmScope.tt_isOpen ).toBeTruthy();
+        elm.trigger('mouseleave');
         expect( elmScope.tt_isOpen ).toBeFalsy();
       }));
     });